chore: fixed project path and name
