+++ NBrzId.Validator.DI.sln
@@ -5,7 +5,11 @@ VisualStudioVersion = 17.11.35327.3
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{D78D34CB-E0F5-43BD-88C9-6CA849BA022D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NBrzId.Validator.DependencyInjection", "NBrzId.Validator.DependecyInjection\NBrzId.Validator.DependencyInjection.csproj", "{90F49CD2-78CB-49B2-936C-001741C49331}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NBrzId.Validator.DependencyInjection", "src\NBrzId.Validator.DependencyInjection\NBrzId.Validator.DependencyInjection.csproj", "{B08D7360-4C99-3EC3-7D1E-622AFF8D68DF}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NBrzId.Validator.DI.UnitTests", "tests\NBrzId.Validator.DI.UnitTests\NBrzId.Validator.DI.UnitTests.csproj", "{05F013C9-16C7-23EB-4BC8-F67F72C08D15}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,16 +17,21 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{90F49CD2-78CB-49B2-936C-001741C49331}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{90F49CD2-78CB-49B2-936C-001741C49331}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{90F49CD2-78CB-49B2-936C-001741C49331}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{90F49CD2-78CB-49B2-936C-001741C49331}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B08D7360-4C99-3EC3-7D1E-622AFF8D68DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B08D7360-4C99-3EC3-7D1E-622AFF8D68DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B08D7360-4C99-3EC3-7D1E-622AFF8D68DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B08D7360-4C99-3EC3-7D1E-622AFF8D68DF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{05F013C9-16C7-23EB-4BC8-F67F72C08D15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{05F013C9-16C7-23EB-4BC8-F67F72C08D15}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{05F013C9-16C7-23EB-4BC8-F67F72C08D15}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{05F013C9-16C7-23EB-4BC8-F67F72C08D15}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{90F49CD2-78CB-49B2-936C-001741C49331} = {D78D34CB-E0F5-43BD-88C9-6CA849BA022D}
+		{B08D7360-4C99-3EC3-7D1E-622AFF8D68DF} = {D78D34CB-E0F5-43BD-88C9-6CA849BA022D}
+		{05F013C9-16C7-23EB-4BC8-F67F72C08D15} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {00068957-46B6-4854-9441-ED63C2523B71}
