+++ src/NBrzId.Validator.DependencyInjection/NBrzId.Validator.DependencyInjection.csproj
@@ -1,0 +1,26 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <BaseOutputPath>bin\</BaseOutputPath>
+    <BaseIntermediateOutputPath>obj\</BaseIntermediateOutputPath>
+    <PackageId>NBrzId.Validator.DependencyInjection</PackageId>
+    <Authors>Christian Martins</Authors>
+    <Description>Contains dependency injection configuration methods for types defined in the package NBrzId.Validator</Description>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <RepositoryUrl>https://github.com/Preulo123/NBrzId.Validator.DI</RepositoryUrl>
+    <PackageIcon>icon.png</PackageIcon>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.5" />
+    <PackageReference Include="NBrzId.Validator" Version="[0.1.0, 1.1.0)" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="icon.png" Pack="true" PackagePath="/" />
+  </ItemGroup>
+
+</Project>
+
+
